Added spacecraftEntity
