+++ src/main/java/io/github/teampropulsive/space/rocket/RocketEntity.java
@@ -1,5 +1,6 @@
 package io.github.teampropulsive.space.rocket;
 
+import io.github.teampropulsive.space.spacecraft.SpacecraftEntity;
 import io.github.teampropulsive.types.Planet;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.attribute.EntityAttributes;
@@ -15,58 +16,15 @@ import org.joml.Vector3f;
 import static io.github.teampropulsive.Propulsive.SPACE;
 import static io.github.teampropulsive.Propulsive.TICKABLE_PLANETS;
 
-public class RocketEntity extends AbstractHorseEntity {
+public class RocketEntity extends SpacecraftEntity {
 
     public boolean hasWarpAbility = false;
     public Vec3d velocity;
+    public Vec3d dockingPosition;
     public float storedOxygen = 0;
-
-    public RocketEntity(EntityType<? extends PathAwareEntity> entityType, World world) {
-        super((EntityType<? extends AbstractHorseEntity>) entityType, world);
-    }
-    @Override
-    public boolean canBeSaddled() {
-        return false;
-    }
-    @Override
-    public boolean isSaddled() {
-        return true;
-    }
-    @Override
-    public EntityView method_48926() {
-        return null;
-    }
-
-    @Override
-    public boolean isAiDisabled() {
-        return true;
-    }
-
-    @Override
-    public void tick() {
-        super.tick();
-    }
-
-    @Override
-    public boolean isOnGround() { // it's on the ground I swear!
-        return true;
-    }
-    @Override
-    protected Vec3d getControlledMovementInput(PlayerEntity controllingPlayer, Vec3d movementInput) {
-        Vec3d a = controllingPlayer.getRotationVecClient();
-        float x = (float) a.x;
-        float y = (float) a.y;
-        float z = (float) a.z;
-        return new Vec3d(x, y, z);
-    }
-    @Override
-    protected float getSaddledSpeed(PlayerEntity controllingPlayer) {
-        return (float)this.getAttributeValue(EntityAttributes.GENERIC_MOVEMENT_SPEED);
-    }
-
-    @Override
-    public boolean isTame() {
-        return true;
+    public float maxOxygen = 1000; // We can settle on a number here later
+    public RocketEntity(EntityType<? extends SpacecraftEntity> entityType, World world) {
+        super(entityType, world);
     }
 
     @Override
@@ -79,21 +37,17 @@ public class RocketEntity extends AbstractHorseEntity {
         //super.jump(strength, movementInput);
     }
 
+    @Override
+    protected Vec3d getControlledMovementInput(PlayerEntity controllingPlayer, Vec3d movementInput) {
+        Vec3d a = controllingPlayer.getRotationVecClient();
+        float x = (float) a.x;
+        float y = (float) a.y;
+        float z = (float) a.z;
+        return new Vec3d(x, y, z);
+    }
 
-    protected Vec3d calculateGravity() {
-        Vec3d velocity = Vec3d.ZERO;
-        if (this.getWorld().getRegistryKey() == SPACE) {
-            for (Planet planet : TICKABLE_PLANETS) {
-                Vec3d distanceDirection = planet.currentPos.subtract(this.getPos());
-                velocity.add(
-                        new Vec3d(
-                                distanceDirection.x * planet.planetSize,
-                                distanceDirection.y * planet.planetSize,
-                                distanceDirection.z * planet.planetSize
-                        )
-                );
-            }
-        }
-        return velocity;
+    @Override
+    protected float getSaddledSpeed(PlayerEntity controllingPlayer) {
+        return (float)this.getAttributeValue(EntityAttributes.GENERIC_MOVEMENT_SPEED);
     }
 }
