+++ src/main/java/io/github/teampropulsive/space/station/StationEntity.java
@@ -1,0 +1,4 @@
+package io.github.teampropulsive.space.station;
+
+public class StationEntity {
+}
